+++ cmake/GoogleCloudCppFeatures.cmake
@@ -97,6 +97,7 @@ set(GOOGLE_CLOUD_CPP_GA_LIBRARIES
     "monitoring"
     "networkconnectivity"
     "networkmanagement"
+    "networksecurity"
     "networkservices"
     "notebooks"
     "optimization"